+++ Lost-Project/ComponenetStore.h
@@ -8,13 +8,13 @@
  
 template<typename T>
 class ComponentStore {
-    std::unordered_map<uint8_t, T> data;
+    std::unordered_map<uint16_t, T> data;
 
 public:
-    T& add(uint8_t id) { data[id] = T(); return data[id]; }
-    bool has(uint8_t id) { return data.count(id) > 0; }
-    T& get(uint8_t id) { return data[id]; }
-    std::pair<uint8_t, T&> getFirst() { return { data.begin()->first, data.begin()->second }; }
+    T& add(uint16_t id) { data[id] = T(); return data[id]; }
+    bool has(uint16_t id) { return data.count(id) > 0; }
+    T& get(uint16_t id) { return data[id]; }
+    std::pair<uint16_t, T&> getFirst() { return { data.begin()->first, data.begin()->second }; }
 };
 
 #endif
