+++ Lost-Project/EntityManager.h
@@ -9,7 +9,7 @@ class EntityManager
 
 public:
 	std::vector<Entity> entities;
-	uint8_t id = 0;
+	uint16_t id = 0;
 	void Init();
 	Entity& CreateEntity();
 	void Cleanup();
